toggling between all and filtered donors for confidence intervals
